+++ lib/qrencoder/version.rb
@@ -1,3 +1,3 @@
-class QRCode
+module QREncoder
   VERSION = '1.1.1'
 end
